fair vis - separated download metadata button into own component
